+++ src/encoded/static/components/index.js
@@ -27,6 +27,7 @@ import DocumentView from './item-pages/DocumentView';
 import StaticSectionView from './item-pages/StaticSectionView';
 import SMaHTSubmissionView from './forms/SMaHTSubmissionView';
 import SearchView from './browse/SearchView';
+import BrowseView from './browse/BrowseView';
 import FileView from './item-pages/FileView';
 
 /**
@@ -58,6 +59,10 @@ content_views.register(SMaHTSubmissionView, 'Search', 'add');
 content_views.register(SearchView, 'Search');
 content_views.register(SearchView, 'Search', 'selection');
 content_views.register(SearchView, 'Search', 'multiselect');
+
+content_views.register(BrowseView, 'Browse');
+content_views.register(BrowseView, 'Browse', 'selection');
+content_views.register(BrowseView, 'Browse', 'multiselect');
 
 // Fallback for anything we haven't registered
 content_views.fallback = function () {